Rename module `databases` to `ntumc_gatekeeper`
